+++ CHANGELOG.md
@@ -1,2 +1,7 @@
+## 0.1.1
+Changed api: stream now provides CarData type
+Added rpmStream(int) and speedStream(double)
+Mcu String cleanup
+
 ## 0.0.1
 First Release